docs(deprecation): Add deprecation notice to README
Unfortunately this project is long since dead. There was also uncommited
files residing on my machine before I had to nuke it due to hardware
malfunction. Namely, the database schema.

Perhaps if I ever become a millionaire, I'll have time to revisit this.
If not, this will forever be a little piece of my development
career that will be shelved until Github dies or the innevitable
heat-death of the universe.

R.I.P Aimless streaming platform